Change coordinates to int instead of float and some mire fixes

+++ src/main/java/world/Food.java
@@ -21,7 +21,7 @@ public class Food extends GameObject {
         return "food";
     }
 
-    public boolean extendsTo(float dx, float dy) {
+    public boolean extendsTo(int dx, int dy) {
         return false;
     }
